storage/endpoint: Add a new component `GCStateStorage` and `GCStateProvider` to storage/endpoints as the new storage layer for GC related metadata (#9109)
ref tikv/pd#8978

Add a new component `GCStateStorage` and `GCStateProvider` to storage/endpoints as the new storage layer for GC related metadata.
This is part of the GC Interface Refactor task. The `GCSafePointStorage` is planned to be deprecated, and the GCStateStorage will be the replacement. It's currently not used, and will be used when we refactor the upper layer (`GCSafePointManager`->`GCStateManager`, the gRPC layer, etc.).

Signed-off-by: MyonKeminta <MyonKeminta@users.noreply.github.com>